- renaming quadrad to quad  - moving quad code to extra file
