+++ src/shapes.rs
@@ -1,1 +1,2 @@
 pub mod shape;
+pub mod quad;
